Refactor protocol version module into focused submodules
